Create QUSHGRProject.ino
Quad Ultrasonic Sensor Hand Gesture Recognition Project
* Currently using 4 Ultrasonic Sensors HC-SR04.
* Detecting hand gestures separately from the top and bottom set of sensors.
* Managing situations where both top and bottom sensors detect gestures simultaneously(*).

(*) It is not exactly simutaneously. but within a period of 30ms. I have to extend this period to ~200ms because now some hand gestures seem to be simutaniously passing in front of both sets of sensors, but they are recognised as seperate gestures.

--> Also needs deletion of old-useless parts of code, addition of comments and "Serial.write()" commands to send data to Unity.