Update qTUSHGRProject.ino
* Added comments.
* Removed unused declared variables.
* Added functionality to the two LEDs:
   * flagLed switches on every time flagL or flagR become equal to 1, after they have been equal to 0.
   * flashingLed switches on every time a gesture to-the-left or to-the-right is recognised.

___
It recognises the most of the gestures correctly, it is not perfect though. I may need to do minor changes to the angle of the sensors, and some parts of the code in the future.

It recognises the most of the gestures correctly, it is not perfect though. I may need do minor changes to the angle of the sensors, and some parts of the code in the future.

It recognises the most of the gestures correctly, it is not perfect though. I may need do minor changes to the angle of the sensors, and some parts of the code in the future.

When I started this project I thought I would compare the distances from each sensor to recognize the gestures. In the end however, gesture recognition is the result of the comparison of the time at which something was detected by the left and right sensors.
Finding a way to finaly take distances into consideration should improve accuracy.